Create Direct and actor

+++ src/main/java/org/example/service/FilmService.java
@@ -1,14 +1,14 @@
 package org.example.service;
 
 import java.util.List;
-import org.example.model.Film;
+import org.example.model.db.Film;
 
 public interface FilmService {
     void create(Film film);
 
     Film getByTitle(String title);
 
-    Film get(int id);
+    Film get(long id);
 
     List<Film> getAll();  // Новый метод
 }
